requirements.txt
pytest
google-cloud-storage
google-cloud-vision
